Fix compile error in c++20 (#510)
In c++20, more constraints are applied:

* constructor and destructor are not allowed to have template id
* adding two enum value needs explicit convertion

(cherry picked from commit 1e70fdc79610782383a5d3495fb3ba851082fbcb)

+++ src/backend/gporca/server/src/unittest/gpopt/minidump/CICGTest.cpp
@@ -187,7 +187,7 @@ CICGTest::EresUnittest_RunUnsupportedMinidumpTests()
 						true /*value*/);
 
 	CAutoTraceFlag atf2(
-		EopttraceDisableXformBase + CXform::ExfDynamicGet2DynamicTableScan,
+		EopttraceDisableXformBase + static_cast<int>(CXform::ExfDynamicGet2DynamicTableScan),
 		true);
 
 	CAutoMemoryPool amp(CAutoMemoryPool::ElcNone);
